fix: address review feedback from @Copilot
Break the status-dot.tsx <-> status-dot-tip.tsx import cycle by extracting
dotLabel (and its SHAPE_LABEL / PR_SHAPE_LABEL maps) into a shared
status-dot-label.ts that both components import. status-dot.tsx re-exports
dotLabel to keep its public surface stable.
